+++ backend/core/domain/transaction.go
@@ -26,8 +26,9 @@ type Transaction struct {
 type TransactionResponse struct {
 	Transaction
 
-	RealizedPnl  float64 `json:"realized_pnl"`
-	PricePerUnit float64 `json:"price_per_unit"`
+	RealizedPnl    float64 `json:"realized_pnl"`
+	EntryPriceUnit float64 `json:"entry_price_unit"`
+	SellPriceUnit  float64 `json:"sell_price_unit"`
 }
 
 type TransactionUpdateReq struct {