+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2026-04-21
+### Changed
+- Splitted `price_per_unit` to 2 mode (`entry_price_unit` (for entry avg price) & `sell_price_unit` (for sell avg price)) to ensure consistency.
+- Optimized portfolio and its total equity retrieval.
+- Refactored backend code.
+- Added more data to export profile.
+- Enhanced `excel.BuildTable` result.
+
 ## [0.2.1] - 2026-04-14
 ### Added
 - Update for Export Profile to Excel feature, now it also includes transaction sheet and packed into table.